Ignore package.el's insistance on initializing in init.el
The package system is configured in configuration.org then
initialized. Package.el seems to, for no good reason, insist that it
is initialized in the init.el, before the package system is properly
configured. This change tricks package.el into thinking that it has
already mucked with the init.el so it leaves it alone on startup. This
allows the configuration.org to properly configure and initialize the
package system.