Add floobits package
It seems that using emacs while developing ruby makes me a bit of a
"black sheep" and less-than-seamless to pair with my peers. I don't
mind /trying/ to use VIM, sublime, etc. in those situations, but I'm
definitely not as proficient with that particular tool. At the
suggesetion of Elle, I'm trying out https://floobits.com -- a service
for cross-editor calibration -- to cut down some of the editor barrier
when I pair.