Try using a light theme
Someone on the internet said "Human eyeballs evolved to identify
silhouettes against bright backgrounds. The whole point of camoflauge
is to break up the human-shaped silhouette to confuse the enemy's
eyeballs. So give light themes a chance." I've been using dark themes
for years, but I figure going to the light-side might worth a try.